polish unittest and refactor the structure of localization
