+++ mmaction/core/evaluation/__init__.py
@@ -1,17 +1,27 @@
-from .accuracy import (average_precision_at_temporal_iou,
+from .accuracy import (MeanClsAcc, TopKAcc, average_precision_at_temporal_iou,
                        average_recall_at_avg_proposals, confusion_matrix,
                        get_weighted_score, interpolated_precision_recall,
                        mean_average_precision, mean_class_accuracy,
                        mmit_mean_average_precision, pairwise_temporal_iou,
                        softmax, top_k_accuracy)
+from .base import BaseMetrics
 from .eval_detection import ActivityNetLocalization
 from .eval_hooks import DistEpochEvalHook, EpochEvalHook
+from .mean_ap import (MeanAP, TemporalMeanAP,
+                      average_precision_at_temporal_iou, eval_ap,
+                      interpolated_precision_recall, mean_average_precision,
+                      mmit_mean_average_precision)
+from .overlaps import pairwise_temporal_iou, temporal_iop, temporal_iou
+from .recall import (ARAN, average_recall_at_avg_proposals,
+                     binary_precision_recall_curve)
 
 __all__ = [
     'DistEpochEvalHook', 'EpochEvalHook', 'top_k_accuracy',
     'mean_class_accuracy', 'confusion_matrix', 'mean_average_precision',
     'get_weighted_score', 'average_recall_at_avg_proposals',
     'pairwise_temporal_iou', 'average_precision_at_temporal_iou',
-    'ActivityNetLocalization', 'softmax', 'interpolated_precision_recall',
-    'mmit_mean_average_precision'
+    'ActivityNetLocalization', 'interpolated_precision_recall',
+    'mmit_mean_average_precision', 'TopKAcc', 'MeanClsAcc', 'BaseMetrics',
+    'average_precision_at_temporal_iou', 'eval_ap', 'TemporalMeanAP', 'MeanAP',
+    'binary_precision_recall_curve', 'ARAN', 'temporal_iop'
 ]
