+++ mmaction/core/evaluation/base.py
@@ -18,12 +18,14 @@ class BaseMetrics(metaclass=ABCMeta):
         """wrap up the results to a dict."""
 
     def print_log_msg(self, log_msg):
+        if self.logger is None:
+            return
         msg = ''
         if isinstance(log_msg, str):
             msg = log_msg
         elif isinstance(log_msg, (tuple, list)):
             msg = ''.join(log_msg)
         elif isinstance(log_msg, dict):
-            for k, v in log_msg:
+            for k, v in log_msg.items():
                 msg += f'\n{k}\t{v:.4f}'
         print_log(msg, logger=self.logger)
