--- mmaction/localization/__init__.py
@@ -1,10 +1,0 @@
-from .bsn_utils import generate_bsp_feature, generate_candidate_proposals
-from .proposal_utils import soft_nms, temporal_iop, temporal_iou
-from .ssn_utils import (eval_ap, load_localize_proposal_file,
-                        perform_regression, temporal_nms)
-
-__all__ = [
-    'generate_candidate_proposals', 'generate_bsp_feature', 'temporal_iop',
-    'temporal_iou', 'soft_nms', 'load_localize_proposal_file',
-    'perform_regression', 'temporal_nms', 'eval_ap'
-]
