+++ mmaction/models/localizers/bmn.py
@@ -4,9 +4,8 @@ import numpy as np
 import torch
 import torch.nn as nn
 
-from ...localization import temporal_iop, temporal_iou
-from ..builder import build_loss
-from ..registry import LOCALIZERS
+from mmaction.core import temporal_iop, temporal_iou
+from mmaction.models.builder import LOCALIZERS, build_loss
 from .base import BaseLocalizer
 from .utils import post_processing
