+++ mmaction/models/localizers/bsn.py
@@ -3,9 +3,8 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
-from ...localization import temporal_iop
-from ..builder import build_loss
-from ..registry import LOCALIZERS
+from mmaction.core import temporal_iop
+from mmaction.models.builder import LOCALIZERS, build_loss
 from .base import BaseLocalizer
 from .utils import post_processing
