+++ mmaction/models/localizers/utils/__init__.py
@@ -1,3 +1,11 @@
+from .bsn_utils import generate_bsp_feature, generate_candidate_proposals
 from .post_processing import post_processing
+from .proposal_utils import soft_nms
+from .ssn_utils import (load_localize_proposal_file, perform_regression,
+                        temporal_nms)
 
-__all__ = ['post_processing']
+__all__ = [
+    'post_processing', 'generate_candidate_proposals', 'generate_bsp_feature',
+    'soft_nms', 'load_localize_proposal_file', 'perform_regression',
+    'temporal_nms'
+]
