+++ mmaction/utils/__init__.py
@@ -1,9 +1,9 @@
 from .collect_env import collect_env
 from .gradcam_utils import GradCAM
 from .logger import get_root_logger
-from .misc import get_random_string, get_shm_dir, get_thread_id
+from .misc import get_random_string, get_shm_dir, get_thread_id, softmax
 
 __all__ = [
     'get_root_logger', 'collect_env', 'get_random_string', 'get_thread_id',
-    'get_shm_dir', 'GradCAM'
+    'get_shm_dir', 'GradCAM', 'softmax'
 ]
