+++ mmaction/utils/misc.py
@@ -2,6 +2,8 @@ import ctypes
 import random
 import string
 
+import numpy as np
+
 
 def get_random_string(length=15):
     """Get random string with letters and digits.
@@ -24,3 +26,9 @@ def get_thread_id():
 def get_shm_dir():
     """Get shm dir for temporary usage."""
     return '/dev/shm'
+
+
+def softmax(x, dim=1):
+    """Compute softmax values for each sets of scores in x."""
+    e_x = np.exp(x - np.max(x, axis=dim, keepdims=True))
+    return e_x / e_x.sum(axis=dim, keepdims=True)
