+++ tests/test_localization_utils.py
@@ -4,9 +4,10 @@ import numpy as np
 import pytest
 from numpy.testing import assert_array_almost_equal, assert_array_equal
 
-from mmaction.localization import (generate_bsp_feature,
-                                   generate_candidate_proposals, soft_nms,
-                                   temporal_iop, temporal_iou)
+from mmaction.core import temporal_iop, temporal_iou
+from mmaction.models.localizers.utils import (generate_bsp_feature,
+                                              generate_candidate_proposals,
+                                              soft_nms)
 
 
 def test_temporal_iou():
